Move the various withXConfig utils to Stack.Runners
